split tests and renames
