+++ test.sh
@@ -2,4 +2,4 @@ cd mocks/
 ./build.sh
 cd ..
 ./build.sh
-cargo test -- --nocapture
+cargo test
